Allow running a docker container from an rpi image on a i386 host.
Running an rpi image as a docker container on the server is
possible since we had the qemu-arm-static binary on images,
which handles arm emulation.
However, this was an amd64 binary (since we were running
the image build process on a dev machine with this architecture).
In order to allow a walt server running on a i386 machine,
we now install the package qemu-user-static with arch=i386.
(This is not a problem with amd64 machines since amd64 processors
are able to handle i386 binaries.)