Add ptpd
This adds the ptpd package, the configuration files,
and creates a file /etc/walt/image.conf that declares
that this new optional feature is implemented and how
to enable it.

New features also declared at the server (in a file called
/etc/walt/server.conf) can be enabled.
This allows old images to work on new servers and vice-versa,
with the ability to enable new features only when they are
implemented on both ends.