Handle the new boot procedure and several rpi models
This image works with rpi-b, rpi-b-plus, rpi-2-b and rpi-3-b boards.
A subdirectory containing the appropriate kernel and dtb is provided
at the root of the image.

Thanks to the new boot mechanism and recent kernels (overlayfs included)
there is no more kernel compilation here, we just download the
compiled binaries on the github repository of the raspberry pi
firmware files.